+++ doc/sphinx/changelog.rst
@@ -19,6 +19,8 @@ Changes
 Fix
 ~~~
 
+- Fix missing tests for :cpp:class:`~mppp::rational` hashing (`#29 <https://github.com/bluescarni/mppp/pull/29>`__).
+
 - Fix some MSVC warnings when compiling the tests in release mode (`#28 <https://github.com/bluescarni/mppp/pull/28>`__).
 
 0.4 (2017-07-29)